[DEV-5856] reorg files even more!
