+++ usaspending_api/etl/elasticsearch_loader_helpers/extract_data.py
